{{ cookiecutter.project_slug }}/requirements/base.txt
# Django
django-environ>=0.9.0,<=0.10.0
Pillow>=9.3.0
{%- if cookiecutter.use_channels != "n" %}
channels[daphne]>=3.0
{%- endif  %}
{%- if cookiecutter.use_celery != "n" %}
django-celery-results
django-celery-beat
{%- endif %}
django-redis
{%- if cookiecutter.api_framework == "RestFramework" %}
djangorestframework
django-filter
markdown
djangorestframework-simplejwt
drf-spectacular
django-cors-headers
{%- elif cookiecutter.api_framework == "GrapheneDjango" %}
graphene_django
django-graphql-jwt
{%- elif cookiecutter.api_framework == "DjangoGrpcFramework" %}
djangogrpcframework
djangorestframework
grpcio
grpcio-tools
{%- elif cookiecutter.api_framework == "BasicHTML" %}
django-crispy-forms
crispy-bootstrap5
{%- endif %}
{%- if cookiecutter.use_async != "n" %}
uvicorn
asyncio
{%- endif %}
